+++ packages/undercut/src/pull/operations/nth.js
@@ -1,7 +1,10 @@
 import { assert } from "../../utils/assert.js";
+import { isPositiveOrZero } from "../../utils/language.js";
 
 export function nth(n) {
-	assert(Number.isSafeInteger(n) && n >= 0, `"n" is required, must be an integer >= 0.`);
+	assert(isPositiveOrZero(n), `"n" is required, must be a number >= 0.`);
+
+	n = Math.trunc(n);
 
 	return function* (iterable) {
 		let index = 0;