+++ packages/undercut/src/pull/operations/skip.test.js
@@ -4,11 +4,14 @@ import { skip, skipWhile } from "./skip.js";
 
 test("skip", () => {
 	expect(() => skip()).toThrow();
+	expect(() => skip(-1)).toThrow();
 
 	expect(targetOf(skip(1), [])).toEqual([]);
 	expect(targetOf(skip(5), [1])).toEqual([]);
 	expect(targetOf(skip(0), [1])).toEqual([1]);
+	expect(targetOf(skip(0.5), [1])).toEqual([1]);
 	expect(targetOf(skip(1), [1, 2])).toEqual([2]);
+	expect(targetOf(skip(1.5), [1, 2])).toEqual([2]);
 	expect(targetOf(skip(3), [1, 2, 3, 4, 5])).toEqual([4, 5]);
 });