+++ packages/undercut/src/pull/operations/take.test.js
@@ -4,10 +4,13 @@ import { take, takeWhile } from "./take.js";
 
 test("take", () => {
 	expect(() => take()).toThrow();
+	expect(() => take(-1)).toThrow();
 
 	expect(targetOf(take(1), [])).toEqual([]);
 	expect(targetOf(take(0), [1])).toEqual([]);
+	expect(targetOf(take(0.5), [1])).toEqual([]);
 	expect(targetOf(take(1), [1, 2])).toEqual([1]);
+	expect(targetOf(take(1.5), [1, 2])).toEqual([1]);
 	expect(targetOf(take(5), [1])).toEqual([1]);
 	expect(targetOf(take(3), [1, 2, 3, 4, 5])).toEqual([1, 2, 3]);
 });